api_container/requirements.txt
boto3
Pillow
openai
requests
uvicorn
fastapi
pydantic
python-dotenv==0.19.0
python-multipart
jinja2
fastapi-healthcheck